Fix clk and rst interfaces, improve 7 seg display
The push buttons are active low. Put inverters to have the desired
functionality.

They also turn on the 7 segment display dots when pressed, for a
visual feedback.

The switches now light up the LEDs above them when their output
is logic ON.

Finally, I tested the clk and rst signals with the instructions
in the ROM and it all worked :)